Remade instruments
New subclass for each instrument, with a new main class with their behaviour. Player can now destroy each instrument idividualy wich make it easy to add points. 

Also created a background on the OnePlayer world wich has a visual bug but I'll be working on a fix for it in the future.